Added an implicit conversion to byte for clenliness

+++ x86/src/main/scala/com/scalaAsm/x86/Instructions/AND.scala
@@ -16,7 +16,7 @@ object AND {
   implicit object and1 extends AND_2[r32, rm32] {
     def get(x: r32, y: rm32) = new Instruction {
       val operands = RM(x,y)
-      val opcode = OneOpcode(0x23.toByte)
+      val opcode = OneOpcode(0x23)
       val modRM: Option[AddressingFormSpecifier] = Some(getAddressingForm2(operands))
       val opcodeExtension = None
      }
