+++ x86/src/main/scala/com/scalaAsm/x86/Instructions/JNZ.scala
@@ -16,7 +16,7 @@ object JNZ {
   implicit object jnz1 extends JNZ_1[imm8] {
     def get(x: imm8) = new Instruction {
       val operands = new OneOperand[imm8](x) {}
-      val opcode = OneOpcode(0x75.toByte)
+      val opcode = OneOpcode(0x75)
       val opcodeExtension = None
       val operand1 = x
       val modRM: Option[AddressingFormSpecifier] = None
