+++ x86/src/main/scala/com/scalaAsm/x86/Instructions/PUSH.scala
@@ -31,7 +31,7 @@ object PUSH extends POWLow {
   implicit object push1 extends PUSH_1[r32] {
     def get(x: r32) = new Instruction {
       val operands = O(x)
-      val opcode = OneOpcode((0x50 + x.ID).toByte)
+      val opcode = OneOpcode(0x50 + x.ID)
       val opcodeExtension = None
       val modRM: Option[AddressingFormSpecifier] = None
     }
@@ -41,7 +41,7 @@ object PUSH extends POWLow {
     def get(x: r16) = new Instruction {
       val operands = O(x)
       val opcodeExtension = None
-      val opcode = OneOpcode((0x50 + x.ID).toByte)
+      val opcode = OneOpcode(0x50 + x.ID)
       val modRM: Option[AddressingFormSpecifier] = None
     }
   }
@@ -49,7 +49,7 @@ object PUSH extends POWLow {
   implicit object push4 extends PUSH_1[imm8] {
     def get(x: imm8) = new Instruction {
       val operands = I[imm8](x)
-      val opcode = OneOpcode(0x6A.toByte)
+      val opcode = OneOpcode(0x6A)
       val opcodeExtension = None
       val modRM: Option[AddressingFormSpecifier] = Some(getAddressingForm1(operands))
     }
@@ -58,7 +58,7 @@ object PUSH extends POWLow {
   implicit object push5 extends PUSH_1[imm16] {
     def get(x: imm16) = new Instruction {
       val operands = I[imm16](x)
-      val opcode = OneOpcode(0x68.toByte)
+      val opcode = OneOpcode(0x68)
       val opcodeExtension = None
       val modRM: Option[AddressingFormSpecifier] = Some(getAddressingForm1(operands))
     }
@@ -68,7 +68,7 @@ object PUSH extends POWLow {
     def get(x: CS) = new Instruction {
       val operands = new OneOperand[CS](x) {}
       val opcodeExtension = None
-      val opcode = OneOpcode(0x0E.toByte)
+      val opcode = OneOpcode(0x0E)
       val modRM: Option[AddressingFormSpecifier] = None
     }
   }
